Issue #153: Move static method getDrupalVersion to the Raw drupal context
